refactor: split observation.rs god-file into modules
