+++ src/commands/bench/observation/tests.rs
@@ -1,757 +1,5 @@
-use std::collections::BTreeMap;
-use std::fs;
-use std::path::{Component, Path, PathBuf};
-
-use homeboy::core::artifacts as artifact_links;
-use homeboy::core::engine::run_dir::{self, RunDir};
-use homeboy::core::extension::bench::{
-    BenchDiagnostic, BenchDiagnosticSource, BenchResults, BenchRunWorkflowResult,
-};
-use homeboy::core::git::short_head_revision_at;
-use homeboy::core::observation::{
-    finding_records_from_budget, merge_metadata, ActiveObservation, ArtifactRecord, NewRunRecord,
-    RunStatus,
-};
-use homeboy::core::rig::RigStateSnapshot;
-
-use crate::commands::utils::resource_policy;
-
-use super::BenchRunArgs;
-
-pub(super) struct BenchObservation(ActiveObservation);
-impl BenchObservation {
-    fn run_id(&self) -> &str {
-        self.0.run_id()
-    }
-}
-
-pub(super) struct BenchObservationSummary {
-    pub run_id: String,
-    pub component_id: String,
-    pub rig_id: Option<String>,
-    pub store_path: String,
-}
-pub(super) struct BenchObservationStart<'a> {
-    pub component_id: &'a str,
-    pub component_label: &'a str,
-    pub source_path: &'a Path,
-    pub args: &'a BenchRunArgs,
-    pub selected_scenarios: &'a [String],
-    pub rig_id: Option<&'a str>,
-    pub rig_snapshot: Option<&'a RigStateSnapshot>,
-    pub run_dir: &'a RunDir,
-}
-pub(super) fn start(start: BenchObservationStart<'_>) -> Option<BenchObservation> {
-    let metadata = bench_observation_initial_metadata(
-        start.component_label,
-        start.args,
-        start.selected_scenarios,
-        start.rig_snapshot,
-        start.run_dir,
-    );
-    let observation = ActiveObservation::start_best_effort(
-        NewRunRecord::builder("bench")
-            .component_id(start.component_id)
-            .command(bench_observation_command(
-                start.component_id,
-                start.args,
-                start.rig_id,
-            ))
-            .cwd_path(start.source_path)
-            .current_homeboy_version()
-            .git_sha(short_head_revision_at(start.source_path))
-            .optional_rig_id(start.rig_id)
-            .metadata(metadata.clone())
-            .build(),
-    )
-    .map(BenchObservation);
-    if let Some(observation) = &observation {
-        write_status_file(
-            start.args,
-            start.run_dir,
-            observation,
-            "running",
-            None,
-            None,
-        );
-    }
-    observation
-}
-
-pub(super) fn finish_success(
-    observation: Option<BenchObservation>,
-    workflow: &mut BenchRunWorkflowResult,
-    run_dir: &RunDir,
-) -> Option<BenchObservationSummary> {
-    let observation = observation?;
-
-    record_bench_observation_artifacts(&observation, workflow, run_dir);
-    let metadata =
-        bench_observation_finish_metadata(observation.0.initial_metadata().clone(), workflow);
-    let status = if workflow.exit_code == 0 {
-        RunStatus::Pass
-    } else {
-        RunStatus::Fail
-    };
-    let summary = BenchObservationSummary {
-        run_id: observation.run_id().to_string(),
-        component_id: observation.0.component_id().unwrap_or_default().to_string(),
-        rig_id: observation.0.rig_id().map(str::to_string),
-        store_path: observation.0.store_path(),
-    };
-    write_status_file(
-        &observation.0.initial_metadata().clone(),
-        run_dir,
-        &observation,
-        workflow.status.as_str(),
-        Some(workflow),
-        None,
-    );
-    observation.0.finish(status, Some(metadata));
-    Some(summary)
-}
-
-/// Build the structured persisted-run pointer surfaced on the bench output
-/// envelope. Carries the run id plus the canonical `homeboy runs show` /
-/// `homeboy runs artifacts` follow-up commands so the compact summary and
-/// downstream tools can locate the full evidence (#3257, #3260).
-pub(super) fn persisted_run_pointer(
-    summary: &BenchObservationSummary,
-) -> homeboy::core::extension::bench::BenchPersistedRun {
-    homeboy::core::extension::bench::BenchPersistedRun {
-        run_id: summary.run_id.clone(),
-        component_id: (!summary.component_id.is_empty()).then(|| summary.component_id.clone()),
-        rig_id: summary.rig_id.clone(),
-        show_command: format!("homeboy runs show {}", summary.run_id),
-        artifacts_command: format!("homeboy runs artifacts {}", summary.run_id),
-    }
-}
-
-pub(super) fn history_hints(summary: &BenchObservationSummary) -> Vec<String> {
-    let mut list_command = format!(
-        "homeboy runs list --kind bench --component {}",
-        summary.component_id
-    );
-    if let Some(rig_id) = &summary.rig_id {
-        list_command.push_str(&format!(" --rig {rig_id}"));
-    }
-
-    vec![
-        format!("Persisted benchmark run ID: {}", summary.run_id),
-        format!("View this run: homeboy runs show {}", summary.run_id),
-        format!("List artifacts: homeboy runs artifacts {}", summary.run_id),
-        format!(
-            "Fetch an artifact: homeboy runs artifact get {} <artifact-name> -o <path>",
-            summary.run_id
-        ),
-        format!("List related bench runs: {list_command}"),
-        format!(
-            "Query persisted bench metadata: homeboy runs query --component {} --since 24h",
-            summary.component_id
-        ),
-        format!(
-            "Observation store: {} (debug path; prefer `homeboy runs ...` commands above)",
-            summary.store_path
-        ),
-    ]
-}
-
-pub(super) fn finish_error(
-    observation: Option<BenchObservation>,
-    error: &homeboy::core::Error,
-    run_dir: &RunDir,
-) {
-    let Some(observation) = observation else {
-        return;
-    };
-
-    record_if_exists(
-        &observation,
-        "bench_results",
-        run_dir.step_file(run_dir::files::BENCH_RESULTS),
-    );
-    record_if_exists(
-        &observation,
-        "resource_summary",
-        run_dir.step_file(run_dir::files::RESOURCE_SUMMARY),
-    );
-    record_memory_timeline_artifacts(&observation, run_dir);
-    let metadata = merge_metadata(
-        observation.0.initial_metadata().clone(),
-        serde_json::json!({
-            "observation_status": "error",
-            "error": error.to_string(),
-        }),
-    );
-    write_status_file(
-        &observation.0.initial_metadata().clone(),
-        run_dir,
-        &observation,
-        "error",
-        None,
-        Some(error),
-    );
-    observation.0.finish_error(Some(metadata));
-}
-
-trait BenchStatusTarget {
-    fn status_file(&self) -> Option<PathBuf>;
-}
-
-impl BenchStatusTarget for BenchRunArgs {
-    fn status_file(&self) -> Option<PathBuf> {
-        self.status_file.clone()
-    }
-}
-
-impl BenchStatusTarget for serde_json::Value {
-    fn status_file(&self) -> Option<PathBuf> {
-        self.get("status_file")
-            .and_then(|value| value.as_str())
-            .map(PathBuf::from)
-    }
-}
-
-fn write_status_file<T: BenchStatusTarget>(
-    target: &T,
-    run_dir: &RunDir,
-    observation: &BenchObservation,
-    status: &str,
-    workflow: Option<&BenchRunWorkflowResult>,
-    error: Option<&homeboy::core::Error>,
-) {
-    let Some(path) = target.status_file() else {
-        return;
-    };
-    if let Some(parent) = path
-        .parent()
-        .filter(|parent| !parent.as_os_str().is_empty())
-    {
-        let _ = fs::create_dir_all(parent);
-    }
-    let artifacts = observation
-        .0
-        .store()
-        .list_artifacts(observation.run_id())
-        .unwrap_or_default();
-    let payload = serde_json::json!({
-        "schema": "homeboy/bench-status/v1",
-        "command": "bench.status",
-        "run_id": observation.run_id(),
-        "kind": observation.0.run().kind,
-        "status": status,
-        "updated_at": chrono::Utc::now().to_rfc3339(),
-        "started_at": observation.0.run().started_at,
-        "finished": workflow.is_some() || error.is_some(),
-        "component_id": observation.0.component_id(),
-        "rig_id": observation.0.rig_id(),
-        "homeboy_version": observation.0.run().homeboy_version,
-        "git_sha": observation.0.run().git_sha,
-        "cwd": observation.0.run().cwd,
-        "run_dir": run_dir.path().to_string_lossy().to_string(),
-        "observation_store": observation.0.store_path(),
-        "artifact_count": artifacts.len(),
-        "artifacts": artifacts,
-        "failure": workflow.and_then(|workflow| workflow.failure.as_ref()),
-        "failure_classification": workflow.and_then(|workflow| workflow.results.as_ref()?.failure_classification.as_ref()),
-        "phase_summaries": workflow.and_then(|workflow| workflow.results.as_ref().map(|results| results.phase_summaries.clone())).unwrap_or_default(),
-        "error": error.map(|error| error.to_string()),
-        "exit_code": workflow.map(|workflow| workflow.exit_code),
-        "gate_failures": workflow.map(|workflow| workflow.gate_failures.clone()).unwrap_or_default(),
-        "hints": workflow.and_then(|workflow| workflow.hints.clone()).unwrap_or_default(),
-    });
-    if let Ok(json) = serde_json::to_vec_pretty(&payload) {
-        let _ = fs::write(path, json);
-    }
-}
-
-fn bench_observation_command(
-    component_id: &str,
-    args: &BenchRunArgs,
-    rig_id: Option<&str>,
-) -> String {
-    let mut parts = vec![
-        "homeboy".to_string(),
-        "bench".to_string(),
-        component_id.to_string(),
-    ];
-    if let Some(rig_id) = rig_id {
-        parts.push(format!("--rig={rig_id}"));
-    }
-    if args.iterations != 10 {
-        parts.push(format!("--iterations={}", args.iterations));
-    }
-    if args.runs != 1 {
-        parts.push(format!("--runs={}", args.runs));
-    }
-    if args.concurrency != 1 {
-        parts.push(format!("--concurrency={}", args.concurrency));
-    }
-    parts.join(" ")
-}
-
-fn bench_observation_initial_metadata(
-    component_label: &str,
-    args: &BenchRunArgs,
-    selected_scenarios: &[String],
-    rig_snapshot: Option<&RigStateSnapshot>,
-    run_dir: &RunDir,
-) -> serde_json::Value {
-    let resource_policy = resource_policy::captured_context()
-        .as_ref()
-        .map(resource_policy::ResourcePolicyContext::to_json)
-        .unwrap_or(serde_json::Value::Null);
-    serde_json::json!({
-        "component_label": component_label,
-        "iterations": args.iterations,
-        "warmup_iterations": args.warmup,
-        "runs": args.runs,
-        "concurrency": args.concurrency,
-        "regression_threshold_percent": args.regression_threshold,
-        "baseline": {
-            "baseline": args.baseline_args.baseline,
-            "ignore_baseline": args.baseline_args.ignore_baseline,
-            "ratchet": args.baseline_args.ratchet,
-        },
-        "profile": args.profile,
-        "selected_scenarios": selected_scenarios,
-        "shared_state": args.shared_state.as_ref().map(|path| path.to_string_lossy().to_string()),
-        "status_file": args.status_file.as_ref().map(|path| path.to_string_lossy().to_string()),
-        "run_dir": run_dir.path().to_string_lossy().to_string(),
-        "rig_state": rig_snapshot,
-        "resource_policy": resource_policy,
-    })
-}
-
-fn bench_observation_finish_metadata(
-    initial_metadata: serde_json::Value,
-    workflow: &BenchRunWorkflowResult,
-) -> serde_json::Value {
-    merge_metadata(
-        initial_metadata,
-        serde_json::json!({
-            "observation_status": workflow.status,
-            "exit_code": workflow.exit_code,
-            "gate_failures": workflow.gate_failures,
-            "baseline_status": baseline_status(workflow),
-            "failure": workflow.failure,
-            "failure_classification": workflow.results.as_ref().and_then(|results| results.failure_classification.clone()),
-            "phase_events": workflow.results.as_ref().map(|results| results.phase_events.clone()).unwrap_or_default(),
-            "phase_summaries": workflow.results.as_ref().map(|results| results.phase_summaries.clone()).unwrap_or_default(),
-            "results": workflow.results,
-            "scenario_metrics": workflow.results.as_ref().map(scenario_metric_summaries).unwrap_or_default(),
-        }),
-    )
-}
-
-fn baseline_status(workflow: &BenchRunWorkflowResult) -> Option<&'static str> {
-    workflow.baseline_comparison.as_ref().map(|comparison| {
-        if comparison.regression {
-            "regression"
-        } else if comparison.has_improvements {
-            "improved"
-        } else {
-            "unchanged"
-        }
-    })
-}
-
-fn scenario_metric_summaries(results: &BenchResults) -> Vec<serde_json::Value> {
-    results
-        .scenarios
-        .iter()
-        .map(|scenario| {
-            serde_json::json!({
-                "scenario_id": scenario.id,
-                "iterations": scenario.iterations,
-                "passed": scenario.passed,
-                "metrics": scenario.metrics,
-                "metric_groups": scenario.metric_groups,
-                "timeline_event_count": scenario.timeline.len(),
-                "span_definition_count": scenario.span_definitions.len(),
-                "span_result_count": scenario.span_results.len(),
-                "metadata": scenario.metadata,
-                "memory": scenario.memory,
-                "artifact_count": scenario.artifacts.len(),
-                "run_count": scenario.runs.as_ref().map(Vec::len),
-                "runs_summary": scenario.runs_summary,
-            })
-        })
-        .collect()
-}
-
-fn record_bench_observation_artifacts(
-    observation: &BenchObservation,
-    workflow: &mut BenchRunWorkflowResult,
-    run_dir: &RunDir,
-) {
-    if let Some(results) = workflow.results.as_mut() {
-        workflow
-            .diagnostics
-            .extend(persist_bench_result_artifact_paths(
-                observation,
-                results,
-                run_dir,
-            ));
-        rewrite_bench_results_file(results, run_dir);
-    }
-
-    record_if_exists(
-        observation,
-        "bench_results",
-        run_dir.step_file(run_dir::files::BENCH_RESULTS),
-    );
-    record_if_exists(
-        observation,
-        "resource_summary",
-        run_dir.step_file(run_dir::files::RESOURCE_SUMMARY),
-    );
-    record_memory_timeline_artifacts(observation, run_dir);
-
-    let Some(results) = workflow.results.as_ref() else {
-        return;
-    };
-    observation.0.record_findings(&finding_records_from_budget(
-        observation.run_id(),
-        &results.budget_findings,
-    ));
-}
-
-fn record_if_exists(observation: &BenchObservation, kind: &str, path: PathBuf) {
-    observation.0.record_artifact_if_file(kind, &path);
-}
-
-fn record_memory_timeline_artifacts(observation: &BenchObservation, run_dir: &RunDir) {
-    let Ok(entries) = fs::read_dir(run_dir.path()) else {
-        return;
-    };
-    for entry in entries.flatten() {
-        let path = entry.path();
-        let Some(name) = path.file_name().and_then(|name| name.to_str()) else {
-            continue;
-        };
-        if name.starts_with("bench-memory-timeline")
-            && (name.ends_with(".json") || name.ends_with(".csv"))
-        {
-            observation
-                .0
-                .record_artifact_if_file("bench_memory_timeline", &path);
-        }
-    }
-}
-
-fn persist_bench_result_artifact_paths(
-    observation: &BenchObservation,
-    results: &mut BenchResults,
-    run_dir: &RunDir,
-) -> Vec<BenchDiagnostic> {
-    let mut diagnostics = Vec::new();
-    let shared_state = results
-        .run_metadata
-        .as_ref()
-        .and_then(|metadata| metadata.shared_state.as_deref());
-    for scenario in &mut results.scenarios {
-        for (name, artifact) in &mut scenario.artifacts {
-            if let Some(diagnostic) = persist_bench_artifact(
-                observation,
-                &scenario.id,
-                None,
-                name,
-                artifact,
-                run_dir,
-                shared_state,
-            ) {
-                diagnostics.push(diagnostic);
-            }
-        }
-        if let Some(runs) = &mut scenario.runs {
-            for (run_index, run) in runs.iter_mut().enumerate() {
-                for (name, artifact) in &mut run.artifacts {
-                    if let Some(diagnostic) = persist_bench_artifact(
-                        observation,
-                        &scenario.id,
-                        Some(run_index),
-                        name,
-                        artifact,
-                        run_dir,
-                        shared_state,
-                    ) {
-                        diagnostics.push(diagnostic);
-                    }
-                }
-            }
-        }
-    }
-    diagnostics
-}
-
-fn persist_bench_artifact(
-    observation: &BenchObservation,
-    scenario_id: &str,
-    run_index: Option<usize>,
-    name: &str,
-    artifact: &mut homeboy::core::extension::bench::BenchArtifact,
-    run_dir: &RunDir,
-    shared_state: Option<&str>,
-) -> Option<BenchDiagnostic> {
-    let kind = artifact.kind.clone().unwrap_or_else(|| name.to_string());
-    let metadata = bench_artifact_metadata(scenario_id, run_index, name, artifact);
-
-    if let Some(url) = artifact.url.clone() {
-        return match observation.0.store().record_url_artifact_with_metadata(
-            observation.run_id(),
-            &kind,
-            &url,
-            metadata,
-        ) {
-            Ok(record) => {
-                apply_recorded_bench_artifact_links(scenario_id, run_index, name, artifact, &record)
-            }
-            Err(error) => Some(bench_artifact_diagnostic(
-                scenario_id,
-                run_index,
-                name,
-                "bench_artifact_url_record_failed",
-                format!("failed to record bench URL artifact `{name}`: {error}"),
-                serde_json::json!({ "url": url }),
-            )),
-        };
-    }
-
-    let original_path = artifact.path.clone()?;
-    if bench_artifact_path_is_blocked(&original_path) {
-        return Some(bench_artifact_diagnostic(
-            scenario_id,
-            run_index,
-            name,
-            "bench_artifact_path_blocked",
-            format!("blocked bench artifact path `{original_path}` for `{name}`"),
-            serde_json::json!({ "path": original_path }),
-        ));
-    }
-
-    let path = resolve_bench_artifact_path(&original_path, run_dir, shared_state);
-    let record = if path.is_file() {
-        observation.0.store().record_artifact_with_metadata(
-            observation.run_id(),
-            "bench_artifact",
-            &path,
-            metadata,
-        )
-    } else if path.is_dir() {
-        observation
-            .0
-            .store()
-            .record_directory_artifact_with_metadata(
-                observation.run_id(),
-                "bench_artifact",
-                &path,
-                metadata,
-            )
-    } else {
-        return Some(bench_artifact_diagnostic(
-            scenario_id,
-            run_index,
-            name,
-            "bench_artifact_path_missing",
-            format!("bench artifact path for `{name}` was not found: {original_path}"),
-            serde_json::json!({
-                "path": original_path,
-                "resolved_path": path.to_string_lossy().to_string(),
-            }),
-        ));
-    };
-
-    match record {
-        Ok(record) => {
-            artifact.path = Some(record.path.clone());
-            apply_recorded_bench_artifact_links(scenario_id, run_index, name, artifact, &record)
-        }
-        Err(error) => Some(bench_artifact_diagnostic(
-            scenario_id,
-            run_index,
-            name,
-            "bench_artifact_record_failed",
-            format!("failed to persist bench artifact `{name}`: {error}"),
-            serde_json::json!({
-                "path": original_path,
-                "resolved_path": path.to_string_lossy().to_string(),
-            }),
-        )),
-    }
-}
-
-fn apply_recorded_bench_artifact_links(
-    scenario_id: &str,
-    run_index: Option<usize>,
-    name: &str,
-    artifact: &mut homeboy::core::extension::bench::BenchArtifact,
-    record: &ArtifactRecord,
-) -> Option<BenchDiagnostic> {
-    artifact.observation_artifact_id = Some(record.id.clone());
-    let public_url = artifact_links::public_artifact_url(record)?;
-    artifact.public_url = Some(public_url.clone());
-    artifact.viewer_links = artifact_links::cached_validated_viewer_links(record, &public_url);
-    artifact.viewer_url = artifact.viewer_links.first().map(|link| link.url.clone());
-    let validation = record.metadata_json.get("public_url_validation")?;
-    let reachable = validation
-        .get("reachable")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-    (!reachable).then(|| {
-        let error = validation
-            .get("error")
-            .and_then(serde_json::Value::as_str)
-            .unwrap_or("public artifact URL was not reachable");
-            bench_artifact_diagnostic(
-                scenario_id,
-                run_index,
-                name,
-                "bench_public_artifact_url_unreachable",
-                format!(
-                    "public artifact URL for bench artifact `{name}` is not reachable; viewer links were not published: {error}"
-                ),
-                serde_json::json!({
-                    "url": validation.get("url").cloned().unwrap_or(serde_json::Value::Null),
-                    "status_code": validation.get("status_code").cloned().unwrap_or(serde_json::Value::Null),
-                    "error": validation.get("error").cloned().unwrap_or(serde_json::Value::Null),
-                }),
-            )
-    })
-}
-
-fn bench_artifact_metadata(
-    scenario_id: &str,
-    run_index: Option<usize>,
-    name: &str,
-    artifact: &homeboy::core::extension::bench::BenchArtifact,
-) -> serde_json::Value {
-    serde_json::json!({
-        "source": "bench",
-        "scenario_id": scenario_id,
-        "run_index": run_index,
-        "name": name,
-        "kind": artifact.kind,
-        "label": artifact.label,
-        "original_path": artifact.path,
-        "url": artifact.url,
-        "viewer": artifact.viewer,
-    })
-}
-
-fn bench_artifact_path_is_blocked(path: &str) -> bool {
-    Path::new(path)
-        .components()
-        .any(|component| matches!(component, Component::ParentDir))
-}
-
-fn bench_artifact_diagnostic(
-    scenario_id: &str,
-    run_index: Option<usize>,
-    name: &str,
-    class: &str,
-    message: String,
-    metadata: serde_json::Value,
-) -> BenchDiagnostic {
-    let mut diagnostic_metadata = BTreeMap::new();
-    diagnostic_metadata.insert("artifact_name".to_string(), serde_json::json!(name));
-    if let Some(object) = metadata.as_object() {
-        for (key, value) in object {
-            diagnostic_metadata.insert(key.clone(), value.clone());
-        }
-    }
-    BenchDiagnostic {
-        class: class.to_string(),
-        message: Some(message),
-        severity: None,
-        source: Some(match run_index {
-            Some(run_index) => BenchDiagnosticSource::ScenarioRun {
-                scenario_id: scenario_id.to_string(),
-                run_index,
-            },
-            None => BenchDiagnosticSource::Scenario {
-                scenario_id: scenario_id.to_string(),
-            },
-        }),
-        metadata: diagnostic_metadata,
-    }
-}
-
-fn rewrite_bench_results_file(results: &BenchResults, run_dir: &RunDir) {
-    let Ok(json) = serde_json::to_vec_pretty(results) else {
-        return;
-    };
-    let _ = fs::write(run_dir.step_file(run_dir::files::BENCH_RESULTS), json);
-}
-
-fn resolve_bench_artifact_path(
-    path: &str,
-    run_dir: &RunDir,
-    shared_state: Option<&str>,
-) -> PathBuf {
-    let artifact_path = PathBuf::from(path);
-    if artifact_path.exists() {
-        return artifact_path;
-    }
-    if artifact_path.is_absolute() {
-        if let Some(shared_state_path) = resolve_shared_state_artifact(&artifact_path, shared_state)
-        {
-            return shared_state_path;
-        }
-        if let Some(preserved_path) = resolve_preserved_invocation_artifact(&artifact_path, run_dir)
-        {
-            return preserved_path;
-        }
-        return artifact_path;
-    }
-    let run_dir_path = run_dir.path().join(path);
-    if run_dir_path.exists() {
-        return run_dir_path;
-    }
-    artifact_path
-}
-
-fn resolve_shared_state_artifact(path: &Path, shared_state: Option<&str>) -> Option<PathBuf> {
-    let shared_state = shared_state?;
-    let relative = path.strip_prefix("/bench-shared-state").ok()?;
-    let candidate = Path::new(shared_state).join(relative);
-    if candidate.exists() {
-        return Some(candidate);
-    }
-    None
-}
-
-fn resolve_preserved_invocation_artifact(path: &Path, run_dir: &RunDir) -> Option<PathBuf> {
-    let mut components = path.components().peekable();
-    while let Some(component) = components.next() {
-        let name = component.as_os_str().to_string_lossy();
-        let Some(short_id) = name.strip_suffix(".a") else {
-            continue;
-        };
-
-        let mut preserved = run_dir
-            .path()
-            .join("invocations")
-            .join(format!("inv-{short_id}"))
-            .join("artifacts");
-        for rest in components {
-            preserved.push(rest.as_os_str());
-        }
-        if preserved.exists() {
-            return Some(preserved);
-        }
-        return None;
-    }
-    None
-}
-
-#[cfg(test)]
-#[path = "../../../tests/commands/bench/observation_artifact_test.rs"]
-mod observation_artifact_test;
-
-#[cfg(test)]
-mod tests {
     use std::fs;
+    use std::path::PathBuf;
 
     use homeboy::core::engine::run_dir::{self, RunDir};
     use homeboy::core::extension::bench::artifact::BenchArtifact;
@@ -761,7 +9,7 @@ mod tests {
     use homeboy::core::observation::ObservationStore;
 
     use super::*;
-    use crate::commands::bench::BenchRigOrder;
+    use crate::commands::bench::{BenchRigOrder, BenchRunArgs};
     use crate::commands::doctor::resources::{
         DoctorOutput, LoadSummary, ProcessSummary, ResourceRecommendation, RigLeaseSummary,
     };
@@ -1752,4 +1000,3 @@ mod tests {
             assert!(run.metadata_json["resource_policy"].is_null());
         });
     }
-}
